Dockerfile, refactorizacion back-front y otras features
