+++ .idea/misc.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Black">
-    <option name="sdkName" value="Python 3.12 (PyCharmMiscProject)" />
+    <option name="sdkName" value="Python 3.13 virtualenv at C:\Users\marta\PycharmProjects\GeneradorDeCuentos\venv" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (Prueba)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.13 virtualenv at C:\Users\marta\PycharmProjects\GeneradorDeCuentos\venv" project-jdk-type="Python SDK" />
 </project>
