+++ __main__.py
@@ -1,4 +1,6 @@
 import uvicorn
+from dotenv import load_dotenv
 
 if __name__ == "__main__":
-    uvicorn.run("backend.api:api", host="127.0.0.1", port=8000, reload=True)
+    load_dotenv()
+    uvicorn.run("backend.routes.api:api", host="127.0.0.1", port=8000, reload=True)
